\name{spherstat-internal}
\alias{birth.sp2}
\alias{death.sp2}
\alias{gridmat.nlon}
\alias{in.W.poly}
\alias{mindist.band}
\alias{mindist.polygon}
\alias{polysph.area.W}
\alias{polysph.area.Wr}
\alias{polysph.area.Wr.exact}
\alias{polysph.area.Wr.grid}
\alias{polysph.area.Wr.int}
\alias{runif.band}
\alias{runif.bandcomp}
\alias{runif.polygon}
\alias{runif.quadrangle}
\alias{runif.sphere}
\alias{runif.wedge}
\alias{sphsincos.vert}

\title{
Internal spherstat functions
}
\description{
Internal spherstat functions
}
\details{
These are generally not to be called by the user.
}
\author{
Adrian Baddeley and Tom Lawrence
}
\keyword{internal}
